requirements.txt
meiteitobangla
requests
python-dotenv
flask
flask-cors